requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
boto3==1.22.5
botocore==1.25.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
decorator==5.1.1
dnspython==2.2.1
dogpile.cache==1.1.6
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.0
future-fstrings==1.2.0
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.1.0
networkx==2.6.3
orjson==3.6.7
packaging==21.3
passlib==1.7.4
pbr==5.9.0
Pillow==9.1.0
pip-autoremove==0.10.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyparsing==3.0.8
pytest==7.1.1
pytest-depends==1.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
stevedore==3.5.0
tomli==2.0.1
typing_extensions==4.1.1
ujson==5.2.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
zipp==3.7.0
